fix(editor): optimistic + rollback for PhaseNodePanel + LocationClueAssignPanel (18.4 followup HIGH-2)
