+++ apps/server/internal/domain/editor/service.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/mmp-platform/server/internal/apperror"
 	"github.com/mmp-platform/server/internal/db"
-	"github.com/mmp-platform/server/internal/engine"
 )
 
 const (
@@ -29,8 +28,6 @@ const (
 	MaxLocationsPerMap = 50
 	MaxCluesPerTheme   = 500
 )
-
-var validContentKeyRe = regexp.MustCompile(`^(story|rules|epilogue|role:[a-z0-9_-]{1,50})$`)
 
 // --- Request / Response types ---
 
@@ -181,6 +178,8 @@ func NewService(q *db.Queries, pool *pgxpool.Pool, logger zerolog.Logger) Servic
 		logger: logger.With().Str("domain", "editor").Logger(),
 	}
 }
+
+// --- Theme ops ---
 
 func (s *service) CreateTheme(ctx context.Context, creatorID uuid.UUID, req CreateThemeRequest) (*ThemeResponse, error) {
 	var theme db.Theme
@@ -398,471 +397,7 @@ func (s *service) SubmitForReview(ctx context.Context, userID, themeID uuid.UUID
 	return toThemeResponse(updated), nil
 }
 
-func (s *service) CreateCharacter(ctx context.Context, creatorID, themeID uuid.UUID, req CreateCharacterRequest) (*CharacterResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-
-	char, err := s.q.CreateThemeCharacter(ctx, db.CreateThemeCharacterParams{
-		ThemeID:     themeID,
-		Name:        req.Name,
-		Description: ptrToText(req.Description),
-		ImageUrl:    ptrToText(req.ImageURL),
-		IsCulprit:   req.IsCulprit,
-		SortOrder:   req.SortOrder,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to create character")
-		return nil, apperror.Internal("failed to create character")
-	}
-	return toCharacterResponse(char), nil
-}
-
-func (s *service) UpdateCharacter(ctx context.Context, creatorID, charID uuid.UUID, req UpdateCharacterRequest) (*CharacterResponse, error) {
-	char, err := s.q.GetThemeCharacter(ctx, charID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, apperror.NotFound("character not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get character")
-		return nil, apperror.Internal("failed to get character")
-	}
-	if _, err := s.getOwnedTheme(ctx, creatorID, char.ThemeID); err != nil {
-		return nil, err
-	}
-
-	updated, err := s.q.UpdateThemeCharacter(ctx, db.UpdateThemeCharacterParams{
-		ID:          charID,
-		Name:        req.Name,
-		Description: ptrToText(req.Description),
-		ImageUrl:    ptrToText(req.ImageURL),
-		IsCulprit:   req.IsCulprit,
-		SortOrder:   req.SortOrder,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to update character")
-		return nil, apperror.Internal("failed to update character")
-	}
-	return toCharacterResponse(updated), nil
-}
-
-func (s *service) DeleteCharacter(ctx context.Context, creatorID, charID uuid.UUID) error {
-	char, err := s.q.GetThemeCharacter(ctx, charID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return apperror.NotFound("character not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get character")
-		return apperror.Internal("failed to get character")
-	}
-	if _, err := s.getOwnedTheme(ctx, creatorID, char.ThemeID); err != nil {
-		return err
-	}
-
-	if err := s.q.DeleteThemeCharacter(ctx, charID); err != nil {
-		s.logger.Error().Err(err).Msg("failed to delete character")
-		return apperror.Internal("failed to delete character")
-	}
-	return nil
-}
-
-func (s *service) ListCharacters(ctx context.Context, creatorID, themeID uuid.UUID) ([]CharacterResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-
-	chars, err := s.q.GetThemeCharacters(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to list characters")
-		return nil, apperror.Internal("failed to list characters")
-	}
-	out := make([]CharacterResponse, len(chars))
-	for i, c := range chars {
-		out[i] = *toCharacterResponse(c)
-	}
-	return out, nil
-}
-
-// UpdateConfigJson is implemented in service_config.go to keep the optimistic
-// lock + RFC 9457 extensions logic isolated and to keep this file under the
-// 500-line limit (see docs/plans/2026-04-15-phase-18.4/design.md #2).
-
-// --- Maps ---
-
-func (s *service) ListMaps(ctx context.Context, creatorID, themeID uuid.UUID) ([]MapResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	maps, err := s.q.ListMapsByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to list maps")
-		return nil, apperror.Internal("failed to list maps")
-	}
-	out := make([]MapResponse, len(maps))
-	for i, m := range maps {
-		out[i] = toMapResponse(m)
-	}
-	return out, nil
-}
-
-func (s *service) CreateMap(ctx context.Context, creatorID, themeID uuid.UUID, req CreateMapRequest) (*MapResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	count, err := s.q.CountMapsByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count maps")
-		return nil, apperror.Internal("failed to count maps")
-	}
-	if count >= MaxMapsPerTheme {
-		return nil, apperror.BadRequest(fmt.Sprintf("theme cannot have more than %d maps", MaxMapsPerTheme))
-	}
-	m, err := s.q.CreateMap(ctx, db.CreateMapParams{
-		ThemeID:   themeID,
-		Name:      req.Name,
-		ImageUrl:  ptrToText(req.ImageURL),
-		SortOrder: req.SortOrder,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to create map")
-		return nil, apperror.Internal("failed to create map")
-	}
-	resp := toMapResponse(m)
-	return &resp, nil
-}
-
-func (s *service) UpdateMap(ctx context.Context, creatorID, mapID uuid.UUID, req UpdateMapRequest) (*MapResponse, error) {
-	m, err := s.q.GetMapWithOwner(ctx, db.GetMapWithOwnerParams{ID: mapID, CreatorID: creatorID})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, apperror.NotFound("map not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get map")
-		return nil, apperror.Internal("failed to get map")
-	}
-	updated, err := s.q.UpdateMap(ctx, db.UpdateMapParams{
-		ID:        m.ID,
-		Name:      req.Name,
-		ImageUrl:  ptrToText(req.ImageURL),
-		SortOrder: req.SortOrder,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to update map")
-		return nil, apperror.Internal("failed to update map")
-	}
-	resp := toMapResponse(updated)
-	return &resp, nil
-}
-
-func (s *service) DeleteMap(ctx context.Context, creatorID, mapID uuid.UUID) error {
-	n, err := s.q.DeleteMapWithOwner(ctx, db.DeleteMapWithOwnerParams{ID: mapID, CreatorID: creatorID})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to delete map")
-		return apperror.Internal("failed to delete map")
-	}
-	if n == 0 {
-		return apperror.NotFound("map not found")
-	}
-	return nil
-}
-
-// --- Locations ---
-
-func (s *service) ListLocations(ctx context.Context, creatorID, themeID uuid.UUID) ([]LocationResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	locs, err := s.q.ListLocationsByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to list locations")
-		return nil, apperror.Internal("failed to list locations")
-	}
-	out := make([]LocationResponse, len(locs))
-	for i, l := range locs {
-		out[i] = toLocationResponse(l)
-	}
-	return out, nil
-}
-
-func (s *service) CreateLocation(ctx context.Context, creatorID, themeID, mapID uuid.UUID, req CreateLocationRequest) (*LocationResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	// verify map belongs to the theme via ownership check
-	_, err := s.q.GetMapWithOwner(ctx, db.GetMapWithOwnerParams{ID: mapID, CreatorID: creatorID})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, apperror.NotFound("map not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get map")
-		return nil, apperror.Internal("failed to get map")
-	}
-	count, err := s.q.CountLocationsByMap(ctx, mapID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count locations")
-		return nil, apperror.Internal("failed to count locations")
-	}
-	if count >= MaxLocationsPerMap {
-		return nil, apperror.BadRequest(fmt.Sprintf("map cannot have more than %d locations", MaxLocationsPerMap))
-	}
-	loc, err := s.q.CreateLocation(ctx, db.CreateLocationParams{
-		ThemeID:              themeID,
-		MapID:                mapID,
-		Name:                 req.Name,
-		RestrictedCharacters: ptrToText(req.RestrictedCharacters),
-		SortOrder:            req.SortOrder,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to create location")
-		return nil, apperror.Internal("failed to create location")
-	}
-	resp := toLocationResponse(loc)
-	return &resp, nil
-}
-
-func (s *service) UpdateLocation(ctx context.Context, creatorID, locID uuid.UUID, req UpdateLocationRequest) (*LocationResponse, error) {
-	l, err := s.q.GetLocationWithOwner(ctx, db.GetLocationWithOwnerParams{ID: locID, CreatorID: creatorID})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, apperror.NotFound("location not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get location")
-		return nil, apperror.Internal("failed to get location")
-	}
-	updated, err := s.q.UpdateLocation(ctx, db.UpdateLocationParams{
-		ID:                   l.ID,
-		Name:                 req.Name,
-		RestrictedCharacters: ptrToText(req.RestrictedCharacters),
-		SortOrder:            req.SortOrder,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to update location")
-		return nil, apperror.Internal("failed to update location")
-	}
-	resp := toLocationResponse(updated)
-	return &resp, nil
-}
-
-func (s *service) DeleteLocation(ctx context.Context, creatorID, locID uuid.UUID) error {
-	n, err := s.q.DeleteLocationWithOwner(ctx, db.DeleteLocationWithOwnerParams{ID: locID, CreatorID: creatorID})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to delete location")
-		return apperror.Internal("failed to delete location")
-	}
-	if n == 0 {
-		return apperror.NotFound("location not found")
-	}
-	return nil
-}
-
-// --- Clues ---
-
-func (s *service) ListClues(ctx context.Context, creatorID, themeID uuid.UUID) ([]ClueResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	clues, err := s.q.ListCluesByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to list clues")
-		return nil, apperror.Internal("failed to list clues")
-	}
-	out := make([]ClueResponse, len(clues))
-	for i, c := range clues {
-		out[i] = toClueResponse(c)
-	}
-	return out, nil
-}
-
-func (s *service) CreateClue(ctx context.Context, creatorID, themeID uuid.UUID, req CreateClueRequest) (*ClueResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	count, err := s.q.CountCluesByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count clues")
-		return nil, apperror.Internal("failed to count clues")
-	}
-	if count >= MaxCluesPerTheme {
-		return nil, apperror.BadRequest(fmt.Sprintf("theme cannot have more than %d clues", MaxCluesPerTheme))
-	}
-	clue, err := s.q.CreateClue(ctx, db.CreateClueParams{
-		ThemeID:     themeID,
-		LocationID:  uuidPtrToPgtype(req.LocationID),
-		Name:        req.Name,
-		Description: ptrToText(req.Description),
-		ImageUrl:    ptrToText(req.ImageURL),
-		IsCommon:    req.IsCommon,
-		Level:       req.Level,
-		ClueType:    req.ClueType,
-		SortOrder:   req.SortOrder,
-		IsUsable:    req.IsUsable,
-		UseEffect:   ptrToText(req.UseEffect),
-		UseTarget:   ptrToText(req.UseTarget),
-		UseConsumed: req.UseConsumed,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to create clue")
-		return nil, apperror.Internal("failed to create clue")
-	}
-	resp := toClueResponse(clue)
-	return &resp, nil
-}
-
-func (s *service) UpdateClue(ctx context.Context, creatorID, clueID uuid.UUID, req UpdateClueRequest) (*ClueResponse, error) {
-	c, err := s.q.GetClueWithOwner(ctx, db.GetClueWithOwnerParams{ID: clueID, CreatorID: creatorID})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, apperror.NotFound("clue not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get clue")
-		return nil, apperror.Internal("failed to get clue")
-	}
-	updated, err := s.q.UpdateClue(ctx, db.UpdateClueParams{
-		ID:          c.ID,
-		LocationID:  uuidPtrToPgtype(req.LocationID),
-		Name:        req.Name,
-		Description: ptrToText(req.Description),
-		ImageUrl:    ptrToText(req.ImageURL),
-		IsCommon:    req.IsCommon,
-		Level:       req.Level,
-		ClueType:    req.ClueType,
-		SortOrder:   req.SortOrder,
-		IsUsable:    req.IsUsable,
-		UseEffect:   ptrToText(req.UseEffect),
-		UseTarget:   ptrToText(req.UseTarget),
-		UseConsumed: req.UseConsumed,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to update clue")
-		return nil, apperror.Internal("failed to update clue")
-	}
-	resp := toClueResponse(updated)
-	return &resp, nil
-}
-
-func (s *service) DeleteClue(ctx context.Context, creatorID, clueID uuid.UUID) error {
-	n, err := s.q.DeleteClueWithOwner(ctx, db.DeleteClueWithOwnerParams{ID: clueID, CreatorID: creatorID})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to delete clue")
-		return apperror.Internal("failed to delete clue")
-	}
-	if n == 0 {
-		return apperror.NotFound("clue not found")
-	}
-	return nil
-}
-
-// --- Contents ---
-
-func (s *service) GetContent(ctx context.Context, creatorID, themeID uuid.UUID, key string) (*ContentResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	if !validContentKeyRe.MatchString(key) {
-		return nil, apperror.BadRequest("invalid content key format")
-	}
-	content, err := s.q.GetContent(ctx, db.GetContentParams{ThemeID: themeID, Key: key})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return &ContentResponse{ThemeID: themeID, Key: key, Body: ""}, nil
-		}
-		s.logger.Error().Err(err).Msg("failed to get content")
-		return nil, apperror.Internal("failed to get content")
-	}
-	resp := toContentResponse(content)
-	return &resp, nil
-}
-
-func (s *service) UpsertContent(ctx context.Context, creatorID, themeID uuid.UUID, key string, body string) (*ContentResponse, error) {
-	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
-		return nil, err
-	}
-	if !validContentKeyRe.MatchString(key) {
-		return nil, apperror.BadRequest("invalid content key format")
-	}
-	content, err := s.q.UpsertContent(ctx, db.UpsertContentParams{
-		ThemeID: themeID,
-		Key:     key,
-		Body:    body,
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to upsert content")
-		return nil, apperror.Internal("failed to upsert content")
-	}
-	resp := toContentResponse(content)
-	return &resp, nil
-}
-
-// --- Validation ---
-
-func (s *service) ValidateTheme(ctx context.Context, creatorID, themeID uuid.UUID) (*ValidationResponse, error) {
-	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
-	if err != nil {
-		return nil, err
-	}
-
-	charCount, err := s.q.CountThemeCharacters(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count characters")
-		return nil, apperror.Internal("failed to validate theme")
-	}
-	mapCount, err := s.q.CountMapsByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count maps")
-		return nil, apperror.Internal("failed to validate theme")
-	}
-	clueCount, err := s.q.CountCluesByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count clues")
-		return nil, apperror.Internal("failed to validate theme")
-	}
-
-	// count locations across all maps
-	locs, err := s.q.ListLocationsByTheme(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to list locations")
-		return nil, apperror.Internal("failed to validate theme")
-	}
-	locCount := int64(len(locs))
-
-	var errs []string
-	if charCount < int64(theme.MinPlayers) {
-		errs = append(errs, fmt.Sprintf("최소 %d명의 캐릭터가 필요합니다 (현재 %d명)", theme.MinPlayers, charCount))
-	}
-	if mapCount == 0 {
-		errs = append(errs, "최소 1개의 맵이 필요합니다")
-	}
-
-	// check culprit character exists
-	chars, err := s.q.GetThemeCharacters(ctx, themeID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to get characters")
-		return nil, apperror.Internal("failed to validate theme")
-	}
-	hasCulprit := false
-	for _, c := range chars {
-		if c.IsCulprit {
-			hasCulprit = true
-			break
-		}
-	}
-	if !hasCulprit {
-		errs = append(errs, "범인 캐릭터가 지정되어야 합니다")
-	}
-
-	return &ValidationResponse{
-		Valid:  len(errs) == 0,
-		Errors: errs,
-		Stats: ValidationStats{
-			Characters: int(charCount),
-			Maps:       int(mapCount),
-			Locations:  int(locCount),
-			Clues:      int(clueCount),
-		},
-	}, nil
-}
-
-// --- Helpers ---
+// --- Shared helpers ---
 
 // getOwnedTheme fetches a theme and verifies creator ownership.
 func (s *service) getOwnedTheme(ctx context.Context, creatorID, themeID uuid.UUID) (db.Theme, error) {
@@ -952,105 +487,4 @@ func toThemeResponse(t db.Theme) *ThemeResponse {
 		resp.ReviewedBy = &uid
 	}
 	return resp
-}
-
-func toCharacterResponse(c db.ThemeCharacter) *CharacterResponse {
-	return &CharacterResponse{
-		ID:          c.ID,
-		ThemeID:     c.ThemeID,
-		Name:        c.Name,
-		Description: textToPtr(c.Description),
-		ImageURL:    textToPtr(c.ImageUrl),
-		IsCulprit:   c.IsCulprit,
-		SortOrder:   c.SortOrder,
-	}
-}
-
-func toMapResponse(m db.ThemeMap) MapResponse {
-	return MapResponse{
-		ID:        m.ID,
-		ThemeID:   m.ThemeID,
-		Name:      m.Name,
-		ImageURL:  textToPtr(m.ImageUrl),
-		SortOrder: m.SortOrder,
-		CreatedAt: m.CreatedAt,
-	}
-}
-
-func toLocationResponse(l db.ThemeLocation) LocationResponse {
-	return LocationResponse{
-		ID:                   l.ID,
-		ThemeID:              l.ThemeID,
-		MapID:                l.MapID,
-		Name:                 l.Name,
-		RestrictedCharacters: textToPtr(l.RestrictedCharacters),
-		SortOrder:            l.SortOrder,
-		CreatedAt:            l.CreatedAt,
-	}
-}
-
-func toClueResponse(c db.ThemeClue) ClueResponse {
-	return ClueResponse{
-		ID:          c.ID,
-		ThemeID:     c.ThemeID,
-		LocationID:  pgtypeUUIDToPtr(c.LocationID),
-		Name:        c.Name,
-		Description: textToPtr(c.Description),
-		ImageURL:    textToPtr(c.ImageUrl),
-		IsCommon:    c.IsCommon,
-		Level:       c.Level,
-		ClueType:    c.ClueType,
-		SortOrder:   c.SortOrder,
-		CreatedAt:   c.CreatedAt,
-		IsUsable:    c.IsUsable,
-		UseEffect:   textToPtr(c.UseEffect),
-		UseTarget:   textToPtr(c.UseTarget),
-		UseConsumed: c.UseConsumed,
-	}
-}
-
-func toContentResponse(c db.ThemeContent) ContentResponse {
-	return ContentResponse{
-		ID:        c.ID,
-		ThemeID:   c.ThemeID,
-		Key:       c.Key,
-		Body:      c.Body,
-		UpdatedAt: c.UpdatedAt,
-	}
-}
-
-func uuidPtrToPgtype(u *uuid.UUID) pgtype.UUID {
-	if u == nil {
-		return pgtype.UUID{}
-	}
-	return pgtype.UUID{Bytes: *u, Valid: true}
-}
-
-func pgtypeUUIDToPtr(u pgtype.UUID) *uuid.UUID {
-	if !u.Valid {
-		return nil
-	}
-	id := uuid.UUID(u.Bytes)
-	return &id
-}
-
-// GetModuleSchemas returns JSON Schema for all registered modules that implement ConfigSchema.
-func (s *service) GetModuleSchemas(_ context.Context) (map[string]json.RawMessage, error) {
-	names := engine.RegisteredModules()
-	schemas := make(map[string]json.RawMessage, len(names))
-
-	for _, name := range names {
-		mod, err := engine.CreateModule(name)
-		if err != nil {
-			s.logger.Warn().Str("module", name).Err(err).Msg("failed to create module for schema collection")
-			continue
-		}
-		cs, ok := mod.(engine.ConfigSchema)
-		if !ok {
-			continue
-		}
-		schemas[name] = cs.Schema()
-	}
-
-	return schemas, nil
 }
