+++ apps/server/internal/domain/editor/service_clue.go
@@ -1,0 +1,271 @@
+package editor
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"regexp"
+
+	"github.com/google/uuid"
+	"github.com/jackc/pgx/v5"
+	"github.com/jackc/pgx/v5/pgtype"
+
+	"github.com/mmp-platform/server/internal/apperror"
+	"github.com/mmp-platform/server/internal/db"
+)
+
+var validContentKeyRe = regexp.MustCompile(`^(story|rules|epilogue|role:[a-z0-9_-]{1,50})$`)
+
+// --- Clues ---
+
+func (s *service) ListClues(ctx context.Context, creatorID, themeID uuid.UUID) ([]ClueResponse, error) {
+	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
+		return nil, err
+	}
+	clues, err := s.q.ListCluesByTheme(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to list clues")
+		return nil, apperror.Internal("failed to list clues")
+	}
+	out := make([]ClueResponse, len(clues))
+	for i, c := range clues {
+		out[i] = toClueResponse(c)
+	}
+	return out, nil
+}
+
+func (s *service) CreateClue(ctx context.Context, creatorID, themeID uuid.UUID, req CreateClueRequest) (*ClueResponse, error) {
+	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
+		return nil, err
+	}
+	count, err := s.q.CountCluesByTheme(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to count clues")
+		return nil, apperror.Internal("failed to count clues")
+	}
+	if count >= MaxCluesPerTheme {
+		return nil, apperror.BadRequest(fmt.Sprintf("theme cannot have more than %d clues", MaxCluesPerTheme))
+	}
+	clue, err := s.q.CreateClue(ctx, db.CreateClueParams{
+		ThemeID:     themeID,
+		LocationID:  uuidPtrToPgtype(req.LocationID),
+		Name:        req.Name,
+		Description: ptrToText(req.Description),
+		ImageUrl:    ptrToText(req.ImageURL),
+		IsCommon:    req.IsCommon,
+		Level:       req.Level,
+		ClueType:    req.ClueType,
+		SortOrder:   req.SortOrder,
+		IsUsable:    req.IsUsable,
+		UseEffect:   ptrToText(req.UseEffect),
+		UseTarget:   ptrToText(req.UseTarget),
+		UseConsumed: req.UseConsumed,
+	})
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to create clue")
+		return nil, apperror.Internal("failed to create clue")
+	}
+	resp := toClueResponse(clue)
+	return &resp, nil
+}
+
+func (s *service) UpdateClue(ctx context.Context, creatorID, clueID uuid.UUID, req UpdateClueRequest) (*ClueResponse, error) {
+	c, err := s.q.GetClueWithOwner(ctx, db.GetClueWithOwnerParams{ID: clueID, CreatorID: creatorID})
+	if err != nil {
+		if errors.Is(err, pgx.ErrNoRows) {
+			return nil, apperror.NotFound("clue not found")
+		}
+		s.logger.Error().Err(err).Msg("failed to get clue")
+		return nil, apperror.Internal("failed to get clue")
+	}
+	updated, err := s.q.UpdateClue(ctx, db.UpdateClueParams{
+		ID:          c.ID,
+		LocationID:  uuidPtrToPgtype(req.LocationID),
+		Name:        req.Name,
+		Description: ptrToText(req.Description),
+		ImageUrl:    ptrToText(req.ImageURL),
+		IsCommon:    req.IsCommon,
+		Level:       req.Level,
+		ClueType:    req.ClueType,
+		SortOrder:   req.SortOrder,
+		IsUsable:    req.IsUsable,
+		UseEffect:   ptrToText(req.UseEffect),
+		UseTarget:   ptrToText(req.UseTarget),
+		UseConsumed: req.UseConsumed,
+	})
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to update clue")
+		return nil, apperror.Internal("failed to update clue")
+	}
+	resp := toClueResponse(updated)
+	return &resp, nil
+}
+
+func (s *service) DeleteClue(ctx context.Context, creatorID, clueID uuid.UUID) error {
+	n, err := s.q.DeleteClueWithOwner(ctx, db.DeleteClueWithOwnerParams{ID: clueID, CreatorID: creatorID})
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to delete clue")
+		return apperror.Internal("failed to delete clue")
+	}
+	if n == 0 {
+		return apperror.NotFound("clue not found")
+	}
+	return nil
+}
+
+// --- Contents ---
+
+func (s *service) GetContent(ctx context.Context, creatorID, themeID uuid.UUID, key string) (*ContentResponse, error) {
+	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
+		return nil, err
+	}
+	if !validContentKeyRe.MatchString(key) {
+		return nil, apperror.BadRequest("invalid content key format")
+	}
+	content, err := s.q.GetContent(ctx, db.GetContentParams{ThemeID: themeID, Key: key})
+	if err != nil {
+		if errors.Is(err, pgx.ErrNoRows) {
+			return &ContentResponse{ThemeID: themeID, Key: key, Body: ""}, nil
+		}
+		s.logger.Error().Err(err).Msg("failed to get content")
+		return nil, apperror.Internal("failed to get content")
+	}
+	resp := toContentResponse(content)
+	return &resp, nil
+}
+
+func (s *service) UpsertContent(ctx context.Context, creatorID, themeID uuid.UUID, key string, body string) (*ContentResponse, error) {
+	if _, err := s.getOwnedTheme(ctx, creatorID, themeID); err != nil {
+		return nil, err
+	}
+	if !validContentKeyRe.MatchString(key) {
+		return nil, apperror.BadRequest("invalid content key format")
+	}
+	content, err := s.q.UpsertContent(ctx, db.UpsertContentParams{
+		ThemeID: themeID,
+		Key:     key,
+		Body:    body,
+	})
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to upsert content")
+		return nil, apperror.Internal("failed to upsert content")
+	}
+	resp := toContentResponse(content)
+	return &resp, nil
+}
+
+// --- Validation ---
+
+func (s *service) ValidateTheme(ctx context.Context, creatorID, themeID uuid.UUID) (*ValidationResponse, error) {
+	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
+	if err != nil {
+		return nil, err
+	}
+
+	charCount, err := s.q.CountThemeCharacters(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to count characters")
+		return nil, apperror.Internal("failed to validate theme")
+	}
+	mapCount, err := s.q.CountMapsByTheme(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to count maps")
+		return nil, apperror.Internal("failed to validate theme")
+	}
+	clueCount, err := s.q.CountCluesByTheme(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to count clues")
+		return nil, apperror.Internal("failed to validate theme")
+	}
+
+	// count locations across all maps
+	locs, err := s.q.ListLocationsByTheme(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to list locations")
+		return nil, apperror.Internal("failed to validate theme")
+	}
+	locCount := int64(len(locs))
+
+	var errs []string
+	if charCount < int64(theme.MinPlayers) {
+		errs = append(errs, fmt.Sprintf("최소 %d명의 캐릭터가 필요합니다 (현재 %d명)", theme.MinPlayers, charCount))
+	}
+	if mapCount == 0 {
+		errs = append(errs, "최소 1개의 맵이 필요합니다")
+	}
+
+	// check culprit character exists
+	chars, err := s.q.GetThemeCharacters(ctx, themeID)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to get characters")
+		return nil, apperror.Internal("failed to validate theme")
+	}
+	hasCulprit := false
+	for _, c := range chars {
+		if c.IsCulprit {
+			hasCulprit = true
+			break
+		}
+	}
+	if !hasCulprit {
+		errs = append(errs, "범인 캐릭터가 지정되어야 합니다")
+	}
+
+	return &ValidationResponse{
+		Valid:  len(errs) == 0,
+		Errors: errs,
+		Stats: ValidationStats{
+			Characters: int(charCount),
+			Maps:       int(mapCount),
+			Locations:  int(locCount),
+			Clues:      int(clueCount),
+		},
+	}, nil
+}
+
+// --- Response mappers ---
+
+func toClueResponse(c db.ThemeClue) ClueResponse {
+	return ClueResponse{
+		ID:          c.ID,
+		ThemeID:     c.ThemeID,
+		LocationID:  pgtypeUUIDToPtr(c.LocationID),
+		Name:        c.Name,
+		Description: textToPtr(c.Description),
+		ImageURL:    textToPtr(c.ImageUrl),
+		IsCommon:    c.IsCommon,
+		Level:       c.Level,
+		ClueType:    c.ClueType,
+		SortOrder:   c.SortOrder,
+		CreatedAt:   c.CreatedAt,
+		IsUsable:    c.IsUsable,
+		UseEffect:   textToPtr(c.UseEffect),
+		UseTarget:   textToPtr(c.UseTarget),
+		UseConsumed: c.UseConsumed,
+	}
+}
+
+func toContentResponse(c db.ThemeContent) ContentResponse {
+	return ContentResponse{
+		ID:        c.ID,
+		ThemeID:   c.ThemeID,
+		Key:       c.Key,
+		Body:      c.Body,
+		UpdatedAt: c.UpdatedAt,
+	}
+}
+
+func uuidPtrToPgtype(u *uuid.UUID) pgtype.UUID {
+	if u == nil {
+		return pgtype.UUID{}
+	}
+	return pgtype.UUID{Bytes: *u, Valid: true}
+}
+
+func pgtypeUUIDToPtr(u pgtype.UUID) *uuid.UUID {
+	if !u.Valid {
+		return nil
+	}
+	id := uuid.UUID(u.Bytes)
+	return &id
+}
